requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.13
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
mysql-connector==2.2.9
psycopg2==2.9.3
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
